buy_sell_web/requirements.txt
asgiref==3.2.7
Babel==2.8.0
bitcoin==1.1.39
blockchain==1.4.4
blockcypher==1.0.81
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-bootstrap-modal-forms==1.5.0
django-cors-headers==3.2.1
django-countries==6.1.2
django-crispy-forms==1.9.0
django-live-support==0.1.9
django-phonenumber-field==4.0.0
django-popup-view-field==0.6.1
django-taggit==1.2.0
django-widget-tweaks==1.4.8
enum-compat==0.0.3
future==0.18.2
idna==2.9
MouseInfo==0.1.3
mysqlclient==1.4.6
phonenumbers==8.12.1
Pillow==7.1.1
PyAutoGUI==0.9.50
PyGetWindow==0.0.8
PyMsgBox==1.0.7
PyMySQL==0.9.3
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.1
python3-xlib==0.15
PyTweening==1.0.3
pytz==2019.3
requests==2.23.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
